Don't show when all lines *can* fit on screen.
This way, if all lines can fit on screen, but e.g., ctrl-e or zt is used
to scroll lines off screen, the scrollbar appearance will be as if all
lines were on screen (not shown). This is consistent with the behavior
at the bottom of the document, where if there are scrollbars, scrolling
the bottom line higher than the bottom window row, results in bars being
displayed as if the bottom line was at the bottom window row.